@@ -18,64 +18,65 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/cli": "^7.15.4",
-    "@babel/core": "^7.15.5",
-    "@babel/eslint-parser": "^7.15.4",
+    "@babel/cli": "^7.24.8",
+    "@babel/core": "^7.24.8",
+    "@babel/eslint-parser": "^7.24.8",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.15.4",
-    "@babel/preset-react": "^7.14.5",
-    "@typescript-eslint/eslint-plugin": "^5.54.0",
-    "@typescript-eslint/parser": "^5.54.0",
-    "babel-jest": "^29.4.3",
-    "babel-loader": "^9.1.2",
+    "@babel/preset-env": "^7.24.8",
+    "@babel/preset-react": "^7.24.7",
+    "@typescript-eslint/eslint-plugin": "^7.16.1",
+    "@typescript-eslint/parser": "^7.16.1",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.1.3",
     "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^11.0.0",
+    "copy-webpack-plugin": "^12.0.2",
     "copyfiles": "^2.4.1",
-    "css-loader": "^6.2.0",
-    "eslint": "^8.35.0",
+    "css-loader": "^7.1.2",
+    "eslint": "^8.56.0",
     "eslint-config-react-app": "^7.0.1",
-    "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-jsx-a11y": "6.7.1",
-    "eslint-plugin-react": "7.32.2",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-webpack-plugin": "4.0.0",
+    "eslint-plugin-import": "2.29.1",
+    "eslint-plugin-jsx-a11y": "6.9.0",
+    "eslint-plugin-react": "7.35.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-webpack-plugin": "4.2.0",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.3.2",
-    "jest": "^29.4.3",
-    "jest-html-reporter": "^3.4.1",
-    "jsdoc": "4.0.2",
-    "react-test-renderer": "^18.2.0",
-    "rimraf": "^4.1.2",
-    "style-loader": "^3.2.1",
-    "webpack": "^5.52.0",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.1.0"
+    "html-webpack-plugin": "^5.6.0",
+    "jest": "^29.7.0",
+    "jest-html-reporter": "^3.10.2",
+    "jsdoc": "4.0.3",
+    "react-test-renderer": "^18.3.1",
+    "rimraf": "^6.0.1",
+    "style-loader": "^4.0.0",
+    "webpack": "^5.93.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.4"
   },
   "dependencies": {
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@fontsource/roboto": "^4.5.0",
-    "@mui/icons-material": "^5.11.9",
-    "@mui/lab": "^5.0.0-alpha.120",
-    "@mui/material": "^5.11.10",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@types/leaflet": "^1.7.4",
-    "axios": "^1.3.4",
-    "i18next": "^22.4.10",
-    "i18next-browser-languagedetector": "^7.0.1",
-    "i18next-http-backend": "^2.1.1",
-    "leaflet": "^1.7.1",
-    "loglevel": "^1.7.1",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
+    "@fontsource/roboto": "^5.0.13",
+    "@mui/icons-material": "^5.16.4",
+    "@mui/lab": "^5.0.0-alpha.172",
+    "@mui/material": "^5.16.4",
+    "@reduxjs/toolkit": "^2.2.6",
+    "@types/leaflet": "^1.9.12",
+    "axios": "^1.7.2",
+    "i18next": "^23.12.2",
+    "i18next-browser-languagedetector": "^8.0.0",
+    "i18next-http-backend": "^2.5.2",
+    "leaflet": "^1.9.4",
+    "loglevel": "^1.9.1",
     "material-ui-flags": "^1.2.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-i18next": "^12.2.0",
-    "react-leaflet": "^4.2.0",
-    "react-redux": "^8.0.5",
-    "react-router-dom": "^6.8.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-i18next": "^15.0.0",
+    "react-leaflet": "^4.2.1",
+    "react-leaflet-draw": "^0.20.4",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.25.1",
     "redux-logger": "^3.0.6",
-    "regenerator-runtime": "^0.13.9",
-    "yargs": "^17.1.1"
+    "regenerator-runtime": "^0.14.1",
+    "yargs": "^17.7.2"
   },
   "babel": {
     "presets": [
